+++ app/src/main/java/com/varsha/pawpals/ui/presentation/PawPalsApp.kt
@@ -40,7 +40,6 @@ import com.varsha.pawpals.ui.presentation.article.DetailArticleScreen
 import com.varsha.pawpals.ui.presentation.article.ExploreArticleScreen
 import com.varsha.pawpals.ui.presentation.community.CommunityScreen
 import com.varsha.pawpals.ui.presentation.home.HomeScreen
-import com.varsha.pawpals.ui.presentation.launch.profile.ProfileScreen
 import com.varsha.pawpals.ui.presentation.profile.EditProfileScreen
 import com.varsha.pawpals.ui.presentation.profile.ProfileScreen
 import com.varsha.pawpals.ui.presentation.schedule.ScheduleScreen
