+++ app/src/main/java/com/varsha/pawpals/ui/presentation/profile/ColumnEdit.kt
@@ -1,4 +1,4 @@
-package com.varsha.pawpals.ui.presentation.launch.profile
+package com.varsha.pawpals.ui.presentation.profile
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
