+++ app/src/main/java/com/varsha/pawpals/ui/presentation/profile/ProfileScreen.kt
@@ -1,5 +1,7 @@
-package com.varsha.pawpals.ui.presentation.launch.profile
+package com.varsha.pawpals.ui.presentation.profile
 
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
@@ -7,55 +9,68 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.AccountCircle
+import androidx.compose.material.icons.filled.AddAlert
 import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.material.icons.filled.ArrowBackIosNew
+import androidx.compose.material.icons.filled.ArrowForwardIos
+import androidx.compose.material.icons.filled.Logout
+import androidx.compose.material.icons.filled.Notifications
+import androidx.compose.material.icons.filled.NotificationsNone
 import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
-import androidx.compose.material3.contentColorFor
+import androidx.compose.material3.surfaceColorAtElevation
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.Font
+import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
 import com.varsha.pawpals.DataUser
+import com.varsha.pawpals.R
 import com.varsha.pawpals.model.UserData
+import com.varsha.pawpals.navigation.Screen
 import com.varsha.pawpals.ui.presentation.component.BackIconItem
 import com.varsha.pawpals.ui.theme.PawPalsTheme
+import kotlin.math.round
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-
-fun EditProfileScreen(
-    navController: NavController,
-    //onBackClicked: () -> Unit
-    //profile: UserData
-
+fun ProfileScreen(
+    //profile: DataUser
+    navController: NavController = rememberNavController(),
 ) {
-    val profil = DataUser.User
-
+    val profile = DataUser.User
     Scaffold(
         topBar = {
             Box(
@@ -66,12 +81,13 @@ fun EditProfileScreen(
                 contentAlignment = Alignment.Center
             ) {
                 Text(
-                    text = "Edit Profil",
+                    text = "Profil",
                     fontSize = 24.sp,
                     fontWeight = FontWeight(600),
                     textAlign = TextAlign.Center,
                     color = Color(0xFF973B2C),
                     modifier = Modifier.align(Alignment.Center)
+
                 )
                 BackIconItem(
                     onBackClicked = { /*TODO*/ },
@@ -80,28 +96,40 @@ fun EditProfileScreen(
             }
         }
     ) { contentPadding ->
-        Column(
-            modifier = Modifier,
+        LazyColumn(
+            modifier = Modifier
+                .padding(contentPadding)
+                .fillMaxSize()
+                .background(color = Color(0xFFFBEDEC)),
             verticalArrangement = Arrangement.Top,
             horizontalAlignment = Alignment.CenterHorizontally
         ) {
-            Column(
-                modifier = Modifier
-                    .padding(contentPadding)
-                    .background(color = Color(0xFFFBEDEC)),
-            ) {
-                ManageProfile()
-                LazyColumn (
+            item {
+                Spacer(modifier = Modifier.height(16.dp))
+                Image(
+                    painter = painterResource(id = profile.photo),
+                    contentDescription = "Profil Photo",
                     modifier = Modifier
-                        //.padding(16.dp)
-                        .clip(RoundedCornerShape(16.dp))
-                        //.background(Color(0xFFFBEDEC))
-                        .background(Color.White)
-                ){
-                    item {
-                        ColumnEdit()
-                    }
-                }
+                        .size(200.dp)
+                        .border(
+                            width = 3.dp,
+                            Color(0xFFD05440),
+                            shape = RoundedCornerShape(size = 110.dp)
+                        )
+                        .clip(CircleShape),
+                    contentScale = ContentScale.Crop
+                )
+
+                Spacer(modifier = Modifier.height(16.dp))
+
+                Text(
+                    text = profile.nama,
+                    fontSize = 24.sp,
+                    color = Color.Black,
+                    modifier = Modifier.padding(bottom = 40.dp)
+                )
+                PetRowContent()
+                ProfilContent(navController = navController)
             }
         }
     }
@@ -109,8 +137,8 @@ fun EditProfileScreen(
 
 @Preview(showBackground = true)
 @Composable
-private fun EditProfilPreview() {
+private fun ProfileScreenPreview() {
     PawPalsTheme {
-        EditProfileScreen(navController = rememberNavController())
+        ProfileScreen()
     }
 }
